dev-workspace/dev_workspace/dev_workspace_stack.py
from aws_cdk import RemovalPolicy, Stack
from aws_cdk import aws_s3 as s3
from constructs import Construct
from dev_workspace.buckets import SimpleBucket


class DevWorkspaceStack(Stack):
    def __init__(self, scope: Construct, construct_id: str, config: dict, **kwargs) -> None:
        super().__init__(scope, construct_id, **kwargs)

        SimpleBucket(self, config["BucketName"])
