dev-workspace/requirements.txt
aws-cdk-lib==2.14.0
constructs>=10.0.0,<11.0.0
pyyaml==6.0
